# Speedup HTTP headers lowercase transformation